+++ chris_salch/src/lib_token/lexer_internal.h
@@ -8,7 +8,7 @@ typedef struct LexerInternal LexerInternalType;
 struct LexerInternal {
     const char *buf; // Character buffer
 
-    off_t offset; // current read offset into the buffer
+    size_t offset; // current read offset into the buffer
 
 
 };
